+++ data/data.txt
@@ -1,3 +1,2 @@
 [E] | [X] | hw  | Dec 13 2023 0100 - Dec 13 2023 0100
 [D] | [ ] | return book  | Dec 13 2023 0100
-[E] | [ ] | hw  | Dec 13 2023 0100 - Dec 13 2023 0100
